cosmetic update
I added another detail how episodes are displayed. the newest gets full episode info, weplayer and shownotes, episodes 2-10 only get subtitle and episode info an the last five items of the list only get the title and publicationDate. It is important to use if, elseif and else as only conditionals because usual ones like 'and' or 'or' don't work with loop.index.