Create Gästeliste
One thing I had on my mind for a while now: I'd like to display everyone who appeared on my podcast on one page, provide links to their social media and their projects as well as a list of episodes with every particular guest. You can easily use this to show appreciation of team members and such as well.